# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ramon api",
        "description": "api for games, movies and books",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/game/{id}": {
            "get": {
                "tags": [
                    "Game"
                ],
                "summary": "Get game by index",
                "description": "Returns a game by the given id",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Game"
                ],
                "summary": "Edit game",
                "description": "Updates a game Consuming a Game object",
                "operationId": "edit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update a Game object ID not required",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Game"
                ],
                "summary": "Delete a game",
                "description": "Delete a game with the given id",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/api/movie/{id}": {
            "get": {
                "tags": [
                    "Movie"
                ],
                "summary": "Get movie by index",
                "description": "Returns a movie by the given id",
                "operationId": "showMovies",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Movie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movie"
                    }
                }
            },
            "put": {
                "tags": [
                    "Movie"
                ],
                "summary": "Updates and returns a movie object",
                "description": "Updates and returns a movie",
                "operationId": "editMovie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "update with a Movie object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Movie"
                ],
                "summary": "Delete a movie",
                "description": "Delete a movie with the given id",
                "operationId": "destroyMovie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Movie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/movie": {
            "post": {
                "tags": [
                    "Movie"
                ],
                "summary": "Creates and returns a movie object",
                "description": "Creates and returns a movie",
                "operationId": "createMovie",
                "requestBody": {
                    "description": "update with a Movie object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/movie/giveTag/{id}": {
            "post": {
                "tags": [
                    "Movie"
                ],
                "summary": "add a tag to a movie",
                "description": "add tag",
                "operationId": "tagMovie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Tag object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    }
}